Extract delay into utils
