+++ lib/Site.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-underscore-dangle */
 
 const cheerio = require('cheerio');
+const delay = require('./util/delay');
 const path = require('path');
 const ignore = require('ignore');
 const ejs = require('ejs');
@@ -82,43 +83,6 @@ function setExtension(filename, ext) {
     path.dirname(filename),
     path.basename(filename, path.extname(filename)) + ext,
   );
-}
-
-/**
-* Creates a function that delays invoking `func` until after `wait` milliseconds have elapsed
-* and the running `func` has resolved/rejected.
-* @param func the promise-returning function to delay,
-*        func should take in a single array
-* @param wait the number of milliseconds to delay
-* @returns delayedFunc that takes in a single argument (either an array or a single value)
-*/
-function delay(func, wait) {
-  let context;
-  let pendingArgs = [];
-  let waitingPromise = null;
-  let runningPromise = Promise.resolve();
-
-  return function (arg) {
-    context = this;
-    if (Array.isArray(arg)) {
-      pendingArgs = pendingArgs.concat(arg);
-    } else {
-      pendingArgs.push(arg);
-    }
-
-    if (waitingPromise === null) {
-      waitingPromise = Promise.all([Promise.delay(wait), runningPromise])
-        .finally(() => {
-          runningPromise = waitingPromise || Promise.resolve();
-          waitingPromise = null;
-          const funcPromise = func.apply(context, [pendingArgs]);
-          pendingArgs = [];
-          return funcPromise;
-        });
-    }
-
-    return waitingPromise;
-  };
 }
 
 /**
